Move editing terminal children to separate hook
